refactor(app_project_structure): update prject structure to follow feature first layer
