--- lib/auth/application/auth_state.dart
@@ -1,13 +1,0 @@
-import 'package:ddd_riverpod/auth/domain/auth_failure.dart';
-import 'package:freezed_annotation/freezed_annotation.dart';
-
-part 'auth_state.freezed.dart';
-
-@freezed
-class AuthState with _$AuthState {
-  const AuthState._();
-  const factory AuthState.initial() = _Initial;
-  const factory AuthState.unauthenticated() = _Unauthenticated;
-  const factory AuthState.authenticated() = _Authenticated;
-  const factory AuthState.failure(AuthFailure failure) = _Failure;
-}
