--- lib/auth/domain/i_auth_facade.dart
@@ -1,8 +1,0 @@
-import 'package:dartz/dartz.dart';
-import 'package:ddd_riverpod/auth/domain/auth_failure.dart';
-
-abstract class IAuthFacade {
-  Future<Either<AuthFailure, Unit>> signInWithGoogle();
-  Future<Either<AuthFailure, Unit>> signOut();
-  Future<bool> isSignIn();
-}
