--- lib/chat/domain/i_chat_facade.dart
@@ -1,10 +1,0 @@
-import 'package:ddd_riverpod/chat/domain/chat_channel_entity.dart';
-
-abstract class IChatFacade {
-  Stream<List<ChatChannelEntity>> chatChannelEntityStream(String uuid);
-  Future<void> updateChatChannel(
-    String channelId,
-    Map<String, dynamic> payload,
-  );
-  Future<ChatChannelEntity?> getChatChannel(String channelId);
-}
