--- lib/chat/domain/message/i_message_facade.dart
@@ -1,7 +1,0 @@
-import 'package:ddd_riverpod/chat/domain/message/message_entity.dart';
-
-abstract class IMessageFacade {
-  Stream<List<MessageEntity>> messageStream(String channelId);
-  Future<void> addMessage(MessageEntity message);
-  Future<void> reportMessage(String messageId, String userId);
-}
