+++ lib/core/presentation/routes/app_router.gr.dart
@@ -14,8 +14,8 @@
 import 'package:auto_route/auto_route.dart' as _i3;
 import 'package:flutter/material.dart' as _i4;
 
-import '../../../auth/presentation/sign_in_page.dart' as _i2;
-import '../../../splash/presentation/splash_page.dart' as _i1;
+import '../../../features/auth/presentation/sign_in_page.dart' as _i2;
+import '../../../features/splash/presentation/splash_page.dart' as _i1;
 
 class AppRouter extends _i3.RootStackRouter {
   AppRouter([_i4.GlobalKey<_i4.NavigatorState>? navigatorKey])
