+++ lib/features/auth/domain/auth_failure.dart
