+++ lib/features/chat/shared/utils/chat_utils.dart
