+++ lib/features/profile/domain/user_entity.dart
